requirements.txt
pipecat-ai[livekit,silero]
openai>=1.3.0
deepgram-sdk>=3.2.0
cartesia>=1.0.0
pymongo>=4.6.0
motor>=3.3.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
livekit>=0.10.0
fastapi>=0.104.0